Refactored camel router naming
